(#76) Terms controller/query service/tests
* Add GetById method to TermsController
* Remove TermController
* Add GetById method to ITermsQueryService
* Remove ITermQueryService
* Add GetById method to ESTermsQueryService
* Remove ESTermQueryService
* Consolidate GetById tests into TermsControllerTest
* Remove TermControllerTest
* Rename and relocate TestDataObjects for GetById QueryService tests
* Refactor GetById-specific tests in ESTermsQueryServiceTests
* Closes #76
